API

Index

Base.wait

XRT.XRTWrap.set_arg!

XRT.XRTWrap.set_arg!

XRT.XRTWrap.set_arg!

XRT.get_kernel_info

XRT.get_section_string

XRT.get_system_info

XRT.BOArray

XRT.BOArray

XRT.XRTWrap.Kernel

XRT.XRTWrap.Run

References

XRT.XRTWrap.Kernel — Method

Kernel(device, uuid, name)

Create a new kernel instance using a device, bitstream uuid, and kernel name.

XRT.XRTWrap.Run — Method

Run(kernel, arg1, args; autostart)

Execute a kernel with the given arguments. To automatically start the execution, set autostart to true. Otherwise, the execution has to be explicitly started by calling start(run::Run)

Base.wait — Method

wait(run)

Wait for a given Run object to complete execution. The method will return as soon as the execution is completed.

XRT.XRTWrap.set_arg! — Method

set_arg!(run, index, val)

Set the argument for a kernel at the given index. Note, that this is a thin wrapper to the C++ API, so the indices start at 0!

XRT.XRTWrap.set_arg! — Method

set_arg!(run, index, val)

Set the argument for a kernel at the given index to a given BOArray. Note, that this is a thin wrapper to the C++ API, so the indices start at 0!

XRT.XRTWrap.set_arg! — Method

set_arg!(run, index, val)

Set the argument for a kernel at the given index to a given BO. Note, that this is a thin wrapper to the C++ API, so the indices start at 0!

XRT.BOArray — Method

BOArray(device, userdata, mem; flags)

Array data type usable with XRT. Can be used like BO but supports indexing and automatic alignment of host buffers.

XRT.BOArray — Method

Array data type usable with XRT. Can be used like BO but supports indexing and automatic alignment of host buffers.

XRT.get_kernel_info — Method

get_kernel_info(xclbin_path::String)

Get information about contained kernels, instances, arguments and their register offsets...

XRT.get_section_string — Method

get_section_string(xclbin_path::String, type::SectionType)

Get the specified raw section from the bitstream file xclbin_path.

XRT.get_system_info — Method

get_system_info(xclbin_path::String)

Get information about resource utilization and connectivity

XRT.@prepare_bitstream — Macro

Parse a bitstream and generate functions for the included kernels. The functions will automatically copy all relevant buffers to the FPGA memory, and execute the Kernel.

It is recommended to generate the kernel functions in a separate module like this:

Julia module DummyBitstream using XRT @prepare_bitstream("my_bitstream.xclbin") end

Afterwards, you find the functions for each kernel in the module. To execute the kernel on a specific device, use the device keyword parameter:

Julia DummyBitstream.kernel_name!(args...; device=XRT.Device(0))