Configuration File xrt.ini

The following reference shows a list of all available configuration parameter. Each parameter can be set using:

XRT.XRTWrap.set! — Function

set!(key::AbstractString, value::AbstractString)
set!(key::AbstractString, value::Integer)

Change xrt.ini string value for specified key. Possible keys can be found in XRT.XRTConfiguration module. See Configuration File xrt.ini section of docs for mor information.

Available configuration parameters

Note: The keys and valid values are based on XRT 2.11 or Vitis 21.1. In newer versions, these may no longer be up-to-date. It is always possible to enter your own keys.

Runtime Group

XRT.XRTConfiguration.API_CHECKS — Constant

Enable or disable OpenCL API checks.

Default: true

    [true|false]

XRT.XRTConfiguration.CPU_AFFINITY — Constant

Pin all runtime threads to specified CPUs.

Example: cpu_affinity = {4,5,6}

    [{N,N,…}]

XRT.XRTConfiguration.EXCLUSIVE_CU_CONTEXT — Constant

When setting true OpenCL process holds exclusive access of the Compute Units.

Default: false

    [false|true]

XRT.XRTConfiguration.RUNTIME_LOG — Constant

Specify where the runtime logs are printed.

Default: console

    [null|console|syslog|filename]

XRT.XRTConfiguration.VERBOSITY — Constant

Verbosity level of log messages. Higher number implies more verbosity.

Default: 4

    [0|1|2|3|4|5|6|7]

Debug Group

XRT.XRTConfiguration.APP_DEBUG — Constant

If true, enable xprint and xstatus command during debugging with xgdb.

Default: false

    [true|false]

XRT.XRTConfiguration.CONTINUOUS_TRACE — Constant

Enables the continuous offload of the device data while the application is running. In the event of a crash/hang a trace file will be available to help debugging.

Default: false

    [false|true]

XRT.XRTConfiguration.CONTINUOUS_TRACE_INTERVAL_MS — Constant

Specifies the interval in millisecond to offload the device data in continous trace mode.

Default: 10

    [N]

XRT.XRTConfiguration.DATA_TRANSFER_TRACE — Constant

Enable device-level AXI transfers trace.

Default: off

    [coarse|fine|off]

XRT.XRTConfiguration.LOP_TRACE — Constant

Enables or disables low overhead profiling.

Default: false

    [false|true]

XRT.XRTConfiguration.PROFILE — Constant

Enable or disable OpenCL host code profile. Set true to generate profile_summary report.

Default: false

    [true|false]

XRT.XRTConfiguration.STALL_TRACE — Constant

Specifies type of stalls to be captured in timeline trace report.

Default: off

    [dataflow|memory|pipe|all|off]

XRT.XRTConfiguration.TIMELINE_TRACE — Constant

Enable or disable profile timeline trace. Set true to generate timeline_trace report.

Default: false

    [true|false]

XRT.XRTConfiguration.TRACE_BUFFER_SIZE — Constant

Specifies the size of DDR/HBM memory for storing trace data. This option only applicable in hardware flow. If no unit is given byte is assumed.

Default: 1M

    [N {K|M|G}]

Emulation Group

XRT.XRTConfiguration.ALIVENESS_MESSAGE_INTERVAL — Constant

Specify the interval in seconds that aliveness messages need to be printed.

Default: 300

    [N]

XRT.XRTConfiguration.LAUNCH_WAVEFORM — Constant

Specify how the waveform is saved and displayed during emulation. The kernel needs to be compiled with debug enabled for the waveform to be saved and displayed in the simulator GUI.

Default: off

    [off|batch|gui]

XRT.XRTConfiguration.PRINT_ERRORS_IN_CONSOLE — Constant

Controls the printing of emulation error messages to users console. Emulation error messages are always logged into a file called emulation_debug.log

Default: true

    [true|false]

XRT.XRTConfiguration.PRINT_INFOS_IN_CONSOLE — Constant

Controls the printing of emulation info messages to users console. Emulation info messages are always logged into a file called emulation_debug.log

Default: true

    [true|false]

XRT.XRTConfiguration.PRINT_WARNING_IN_CONSOLE — Constant

Controls the printing of emulation warning messages to users console. Emulation warning messages are always logged into a file called emulation_debug.log

Default: true

    [true|false]

XRT.XRTConfiguration.TIMEOUT_SCALE — Constant

Specify the time scaling unit of timeout specified clPollStreams command, otherwise Emulation does not support timeout specified in clPollStreams command.

Default: na (not applicable)

    [na|ms|sec|min]