Module _XRTInternal

All functions specified in this module are for internal usage and should in general not be called directly!

Index

XRT._XRTInternal.call_utility

XRT._XRTInternal.construct

XRT._XRTInternal.convert_bytes

XRT._XRTInternal.draw_box!

XRT._XRTInternal.draw_kernel_cu_connection!

XRT._XRTInternal.draw_kernel_streams

XRT._XRTInternal.get_info

XRT._XRTInternal.get_info

XRT._XRTInternal.highlight_offline_kwargs!

XRT._XRTInternal.transform_header

XRT._XRTInternal.xrtprint_argslist

XRT._XRTInternal.xrtprint_cu

XRT._XRTInternal.xrtprint_device

XRT._XRTInternal.xrtprint_dynamic_regions

XRT._XRTInternal.xrtprint_electrical

XRT._XRTInternal.xrtprint_mechanical_thermal

XRT._XRTInternal.xrtprint_memory

XRT._XRTInternal.xrtprint_platform

XRT._XRTInternal.xrtprint_simple

XRT._XRTInternal.xrtprint_simple_header

XRT._XRTInternal.xrtprint_xclbin

XRT._XRTInternal.xrtprint_xclbinarg

XRT._XRTInternal.xrtprint_xclbinip

XRT._XRTInternal.xrtprint_xclbinkernel

XRT._XRTInternal.xrtprint_xclbinmem

XRT._XRTInternal.State

XRT._XRTInternal.XilinxDeviceArray

References

XRT._XRTInternal.State — Type

mutable struct State

State is a singleton object that stores global parameters, like devices and emulation mode.

XRT._XRTInternal.XilinxDeviceArray — Type

struct XilinxDeviceArray

A XilinxDeviceArray behaves like a normal Julia Vector{XRT.XilinxDevice}. The difference is, that it always points to the first element, when emulation mode is set.

XRT._XRTInternal.construct — Method

construct()

Constructs and initialises a singleton object of type XRT._XRTInternal.State. It checks for a set emulation mode, and detects and registers available devices.

XRT._XRTInternal.get_info — Method

get_info(
    device::XRT.XRTWrap.Device,
    param::XRT.XRTWrap.DeviceInformationParameters.Type
) -> Any

Fetches the device information based on the given parameter. It returns a value of the corresponding type or nothing if the parameter is not available.

XRT._XRTInternal.get_info — Method

get_info(
    device::XRT.XilinxDevice,
    param::XRT.XRTWrap.DeviceInformationParameters.Type
) -> Any

Fetches the device information based on the given parameter.

XRT._XRTInternal.call_utility — Method

call_utility(
    args::AbstractString...;
    ignorestatus
) -> String

Function used to call an XRT command-line tool from Julia. It detects whether xrt_jll or a native installation is used.

XRT._XRTInternal.draw_box! — Method

draw_box!(
    graph::Matrix{Char},
    x::Int64,
    y::Int64,
    width::Int64,
    title::AbstractString,
    input::AbstractString
)

Internal function that draws a five rows high box with a title and an input string to given coordinates of a char matrix. The function throws an exception if the box does not fit the matrix. The output looks like following:

┌──────────────┐
│    title     │
│              │
│ input string │
└──────────────┘

XRT._XRTInternal.draw_kernel_cu_connection! — Method

draw_kernel_cu_connection!(
    graph::Matrix{Char},
    x::Int64,
    y::Int64,
    kernel_width::Int64,
    cu_connection_width::Int64,
    required_cu_height::Int64
) -> Union{Nothing, Char}

Internal function that draws connecting arrows from the kernel box to its corresponding cu boxes.

XRT._XRTInternal.draw_kernel_streams — Function

draw_kernel_streams(xclbin::Xclbin; ...)
draw_kernel_streams(
    xclbin::Xclbin,
    filter::AbstractString;
    io,
    force_io
)

See XRT.visualize_streams.

Prettyprinting

XRT._XRTInternal.convert_bytes — Method

convert_bytes(
    bytes::Union{AbstractString, Integer}
) -> String

This functions parses a (hexadecimal) string to an integer and transforms it to a human-readable format with Byte units.

julia> _XRTInternal._XRTInternal.convert_bytes("0x10000040")
"256 MB"

XRT._XRTInternal.highlight_offline_kwargs! — Method

highlight_offline_kwargs!(
    kwargs::Dict
) -> PrettyTables.TextFormat

Takes a dict of keyword arguments for the use with pretty_table function. Makes the table outline dotted and greys the complete table.

XRT._XRTInternal.transform_header — Method

transform_header(header::AbstractString) -> String

Takes a string or header obtained from JSON object that is in format "foo_bar". Transforms it to a string as "Foo Bar" and applies some hardcoded exceptions for units and abbreviations.

XRT._XRTInternal.xrtprint_argslist — Method

xrtprint_argslist(args::Vector{XRT.XclbinArg}; io)

Internal print function for a list of XRT.XclbinArg objects in a tabular form.

XRT._XRTInternal.xrtprint_cu — Method

xrtprint_cu(cu::XRT._XRTInternal.PrintComputeUnit; io)

Internal print function for a single PrintComputeUnit object.

XRT._XRTInternal.xrtprint_device — Method

xrtprint_device(device::XRT.XilinxDevice; io)

Internal print function for general XilinxDevice data.

XRT._XRTInternal.xrtprint_dynamic_regions — Method

xrtprint_dynamic_regions(json::LazyJSON.Object; io)

Internal print function for dynamic-regions report.

XRT._XRTInternal.xrtprint_electrical — Method

xrtprint_electrical(json::LazyJSON.Object; io)

Internal print function for electrical report.

XRT._XRTInternal.xrtprint_mechanical_thermal — Method

xrtprint_mechanical_thermal(json::LazyJSON.Object; io)

Internal print function for mechanical and thermal reports.

XRT._XRTInternal.xrtprint_memory — Method

xrtprint_memory(json::LazyJSON.Object; io)

Internal print function for memory report.

XRT._XRTInternal.xrtprint_platform — Method

xrtprint_platform(json::LazyJSON.Object; io)

Internal print function for platform report.

XRT._XRTInternal.xrtprint_simple — Method

xrtprint_simple(json::LazyJSON.Object; io)

Internal print function for reports which just contain strings as values.

XRT._XRTInternal.xrtprint_simple_header — Method

xrtprint_simple_header(title::AbstractString; io)

Internal print function for a section header.

XRT._XRTInternal.xrtprint_xclbin — Method

xrtprint_xclbin(xclbin::Xclbin; io)

Internal print function for XRT.Xclbin objects.

XRT._XRTInternal.xrtprint_xclbinarg — Method

xrtprint_xclbinarg(arg::XRT.XclbinArg; io)

Internal print function for XRT.XclbinArg objects.

XRT._XRTInternal.xrtprint_xclbinip — Method

xrtprint_xclbinip(ip::XRT.XclbinIP; io)

Internal print function for XRT.XclbinIP objects. Does also print compute unit represantation of kernels.

XRT._XRTInternal.xrtprint_xclbinkernel — Method

xrtprint_xclbinkernel(kernel::XRT.XclbinKernel; io)

Internal print function for XRT.XclbinKernel objects.

XRT._XRTInternal.xrtprint_xclbinmem — Method

xrtprint_xclbinmem(mem::XRT.XclbinMem; io)

Internal print function for XRT.XclbinMem objects.